chore: move validate package to datacatalog provider
