+++ cli/internal/providers/datacatalog/validate/refs.go
