+++ cli/internal/providers/datacatalog/validate/refs_test.go
