+++ cli/internal/providers/datacatalog/validate/required_keys.go
