+++ cli/internal/providers/datacatalog/validate/required_keys_test.go
