+++ cli/internal/providers/datacatalog/validate/validate.go
